GEODE-3267: Refactoring QueueCommands
